feat: Customize `include` and `exclude` options
