+++ package.json
@@ -88,6 +88,7 @@
     "@babel/parser": "^7.22.16",
     "@babel/plugin-syntax-jsx": "^7.22.5",
     "@babel/plugin-syntax-typescript": "^7.22.5",
+    "@rollup/pluginutils": "^5.0.5",
     "@vue/compiler-dom": "^3.3.4",
     "magic-string": "^0.30.3",
     "unplugin": "^1.4.0"
