+++ src/core/index.ts
@@ -1,8 +1,9 @@
-import type { UnpluginFactory } from 'unplugin';
-import { createUnplugin } from 'unplugin';
-import type { Options } from '../types';
-import { filter_ID } from './filter_ID';
+import { type UnpluginFactory, createUnplugin } from 'unplugin';
+import { createFilter } from '@rollup/pluginutils';
+import { type ResolvedOptions, type Options } from '../types';
+import { TRACE_ID } from './constants';
 import { transform } from './transform';
+import { parse_ID } from './parse_ID';
 
 export const unpluginFactory: UnpluginFactory<Options> = (options = {}) => {
   if (process.env.NODE_ENV !== 'development') {
@@ -12,23 +13,44 @@ export const unpluginFactory: UnpluginFactory<Options> = (options = {}) => {
   }
 
   const opts = resolveOptions(options);
+  const filter = createFilter(opts.include, opts.exclude);
 
   return {
     name: 'unplugin-vue-source',
     enforce: 'pre',
-    transformInclude: filter_ID,
+    transformInclude(id) {
+      if (filter(id)) {
+        const parsed = parse_ID(id);
+
+        if (parsed.isSfc) {
+          const { query } = parsed;
+          // vue cli | vue-loader
+          if (query.type === 'template') {
+            return true;
+          }
+          return (
+            // vite-plugin-vue
+            !query[TRACE_ID] &&
+            // rollup-plugin-vue
+            !query['rollup-plugin-vue']
+          );
+        }
+
+        return true;
+      }
+    },
     transform(code, id) {
       return transform(code, id, opts);
     },
   };
 };
 
-function resolveOptions(options: Options): Required<Options> {
-  const { root = process.cwd(), sourceMap = false } = options;
-
+function resolveOptions(opts: Options): ResolvedOptions {
   return {
-    root,
-    sourceMap,
+    root: opts.root ?? process.cwd(),
+    sourceMap: opts.sourceMap ?? false,
+    include: opts.include ?? '**/*.{vue,jsx.tsx}',
+    exclude: opts.exclude ?? 'node_modules/**',
   };
 }
