+++ src/types.ts
@@ -11,4 +11,11 @@ export interface Options {
    * @default false
    */
   sourceMap?: boolean;
+
+  /** @default '**\/*.{vue,jsx.tsx}' */
+  include?: string | RegExp | (string | RegExp)[];
+  /** @default 'node_modules/**' */
+  exclude?: string | RegExp | (string | RegExp)[];
 }
+
+export type ResolvedOptions = Required<Options>;
